requirements.txt
# Core Framework
fastapi>=0.115.8
uvicorn>=0.34.0
python-multipart>=0.0.20
pydantic>=2.10.6
pydantic-settings>=2.8.0
email-validator>=2.2.0

# Database
sqlalchemy>=2.0.38
alembic>=1.14.1
psycopg2-binary>=2.9.10
redis>=5.2.1
aioredis>=2.0.1

# Task Queue
celery>=5.4.0
flower>=2.0.1

# Authentication & Security
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.9.0.post0

# HTML Parsing & Analysis
beautifulsoup4>=4.13.3
lxml>=5.3.1
requests>=2.32.3
aiohttp>=3.11.12
html5lib>=1.1

# Performance & Monitoring
prometheus-client>=0.21.1

# Testing
pytest>=8.3.4
pytest-asyncio>=0.25.3
httpx>=0.28.1
pytest-cov>=6.0.0

# Development
black>=25.1.0
isort>=6.0.0
flake8>=7.1.2
mypy>=1.15.0